+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  - add `radioStyle` to be load in the basic style
  - add `checkboxStyle`
  - update `survey-angular-ui` from `1.9.109` to `1.9.125`
+ - add an option to upload files to the server when uploaded with `survey-js`
 
 ### Bugfix
  - #2 - fix the validation, `url` is correctly recognized with the capacitor.